pyproject.toml
[project]
name = "gdg-devfest-graph-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ell-ai>=0.0.14",
    "kuzu>=0.6.1",
    "lancedb>=0.14.0",
    "llama-index>=0.11.19",
    "llama-index-embeddings-openai>=0.2.5",
    "llama-index-graph-stores-kuzu>=0.3.2",
    "llama-index-llms-openai>=0.2.15",
    "llama-index-postprocessor-cohere-rerank>=0.2.1",
    "llama-index-vector-stores-lancedb>=0.2.4",
    "python-dotenv>=1.0.1",
]

[tool.uv]
dev-dependencies = [
    "ipykernel>=6.29.5",
]
